Phase 14.30 (R-000168 compile surface restored): 0-error cargo check on all compile-expected targets
The post-11.1-U x86_64 varargs shims (11.1-Z.2 R-000176) and xslt timing
calibration had silently broken the platform compile-expected evidence:
cargo check --target aarch64-unknown-linux-gnu / i686-unknown-linux-gnu /
armv7-unknown-linux-gnueabihf failed with VaListTag-undefined errors on the
xmlParser{Error,Warning,ValidityError,ValidityWarning}V receivers and an
AtomicI64 fetch_add fed a 32-bit c_long in xsltCalibrateAdjust.

- the four legacy *V receivers and the XML_PARSER_*_SAX1 consts are
  x86_64-gated with the SysV register-save-area shim; non-x86_64 builds get
  raw-message forwarding defaults (the candidate's own raises always hand
  these slots pre-formatted text; C consumers variadically calling the
  DEFAULT handler on other ABIs remain a documented varargs-family
  obligation in PLATFORM_SURFACE_ATLAS)
- xsltCalibrateAdjust widens c_long to the i64 accumulator (no-op on 64-bit)
- re-validated: cargo check --lib for aarch64 / i686 / armv7 / musl all
  0 errors; host x86_64 build + 1254 lib tests green; ELF-VERSIONING court
  20/20; DSO-STATE-COHERENCE PASS (parity)

RESIDUAL_LEDGER: R-000168 stays OPEN by design (runtime execution outside
Linux x86-64 is the CI-matrix obligation) with the refreshed compile
evidence in its history. 1 open residual remains.

+++ src/abi/exports_xslt_util.rs
@@ -518,8 +518,9 @@ pub unsafe extern "C" fn xsltTimestamp() -> c_long {
 #[no_mangle]
 pub unsafe extern "C" fn xsltCalibrateAdjust(delta: c_long) {
     // Upstream stores a process-lifetime calibration offset; the candidate
-    // applies it to the monotonic clock read.
-    CALIBRATION_OFFSET.fetch_add(delta, core::sync::atomic::Ordering::Relaxed);
+    // applies it to the monotonic clock read. `c_long` is 32-bit on 32-bit
+    // ABIs — widen to the i64 accumulator (R-000168 calibration arithmetic).
+    CALIBRATION_OFFSET.fetch_add(delta as i64, core::sync::atomic::Ordering::Relaxed);
 }
 
 static CALIBRATION_OFFSET: core::sync::atomic::AtomicI64 = core::sync::atomic::AtomicI64::new(0);
